requirements.txt
absl-py==1.3.0
altair==4.2.0
astunparse==1.6.3
backports.zoneinfo==0.2.1;python_version<"3.9"
blinker==1.5
cachetools==5.2.0
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
ffmpeg==1.4
ffmpeg-python==0.2.0
flatbuffers==1.12
fonttools==4.38.0
future==0.18.2
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
imageio==2.22.4
joblib==1.2.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
matplotlib==3.6.2
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
pandas==1.5.2
Pillow==9.3.0
protobuf==3.20.3
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pympler==1.0.1
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
semver==2.13.0
streamlit==1.15.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
threadpoolctl==3.1.0
toolz==0.12.0
tzdata==2022.7
tzlocal==4.2
validators==0.20.0
watchdog==2.1.9
Werkzeug==2.2.2
wget==3.2
